🐛 Fix: formatDate 공통유틸함수로 분리
